+++ study/src/test/java/sql/RespondentHobbyDto.java
@@ -1,4 +1,4 @@
-package db;
+package sql;
 
 public class RespondentHobbyDto {
     private final long respondentId;
